internship_project_2025/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.42
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic[email]==2.11.7
pydantic-settings==2.10.1
qrcode==8.2
pillow==11.3.0
fpdf2==2.8.3